app/function/requirements.txt
google-api-core==1.15.0
google-api-python-client==1.7.11
google-auth==1.10.0
google-cloud-core==1.1.0
google-cloud-kms==0.2.0
google-cloud-storage==1.24.1
requests==2.22.0
